+++ src/modules/taskDB.js
@@ -1,0 +1,19 @@
+const todoTasks = [
+    {
+      description: 'Read the documentation',
+      completed: false,
+      index: 1,
+    },
+    {
+      description: 'Create the Repository',
+      completed: false,
+      index: 2,
+    },
+    {
+      description: 'Set the Linters',
+      completed: false,
+      index: 3,
+    },
+  ];
+
+  export default todoTasks
